refactor: static 메서드를 인터페이스로 추상화
- 인터페이스로 추상화 하여 static을 사용하지 않고 객체를 생성하도록 변경
